Linux fix: Search with schemas fails in cold keyrings
When the keyring is cold (e.g. after restart without a login or after
running `gnome-keyring-daemon --replace` on the command line), lookup
using schemas would return null entries without prompting the user for
their passwords.

This is especially dangerrous in situations where the user wants to
write an entry to the keyring. In such case, if there were previously
stored secrets in the FlutterSecretStorage entry, those would be ignored
and only the new entry would be stored. Effectively erasing all
previously stored passwords.

The issue is explained here:
https://gitlab.gnome.org/GNOME/gnome-keyring/-/issues/89

However, the solution used in this commit is the one used in
the chromium browser: http://crbug.com/660005